Add files via upload
The Scenario: The Fynbos Flower company is flexible on the hours that employees work, and as a consequence, timekeeping is an important administrative practice. The company records the dates and times of shifts (for want of a better word). Employees sign-in when they arrive and sign-out when they leave. The start of a shift and the end of a shift always fall within the same week. Though, currently, none of the employees bear the same name, the company assigns each a unique ID number that may be used to disambiguate.

The project is split into three parts: 
Part one  concerns constructing a UML class diagram for the classes, showing attributes,
constructors, methods, and associations.
Part two involves using Week, CalendarTime, Date, Time, Duration and TimeUnit to implement the Shift class.
Part three involves implementing the Employee class.